IDM-VTON/requirements.txt
absl-py==2.1.0
accelerate==0.25.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
av==14.0.1
basicsr==1.4.2
bitsandbytes==0.39.0
CacheControl==0.14.2
cachetools==5.5.0
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
datasets==3.2.0
diffusers==0.25.0
dill==0.3.8
einops==0.7.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.6
ffmpy==0.5.0
firebase-admin==6.6.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==5.10.0
gradio_client==1.5.3
grpcio==1.70.0
grpcio-status==1.69.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.25.2
humanfriendly==10.0
hyperframe==6.1.0
imageio==2.36.1
importlib_metadata==8.5.0
iopath==0.1.10
Jinja2==2.11.3
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.4.1
mpmath==1.2.1
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==2.6.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optimum==1.23.3
optree==0.14.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==9.0.1
platformdirs==4.3.6
portalocker==2.10.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==5.4.1
qdrant-client==1.13.2
regex==2024.11.6
rich==13.9.4
rsa==4.9
ruff==0.8.6
safehttpx==0.1.6
safetensors==0.5.1
scikit-image==0.24.0
scipy==1.11.1
semantic-version==2.10.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.19.0a20250107
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2024.12.12
tokenizers==0.13.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.2.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.30.2
typer==0.15.1
typing_extensions==4.5.0
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.34.0
websockets==14.1
Werkzeug==3.1.3
win-inet-pton==1.1.0
wrapt==1.17.2
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
